+++ webapp/src/main/webapp/src/app/school-grade/results/school-results.component.ts
@@ -11,10 +11,8 @@ import { Grade } from "../grade.model";
 import { isNullOrUndefined } from "util";
 import { Angulartics2 } from "angulartics2";
 import { AssessmentsComponent } from "../../assessments/assessments.component";
-import { Exam } from "../../assessments/model/exam.model";
-import { ExamFilterService } from "../../assessments/filters/exam-filters/exam-filter.service";
-import { CsvBuilder } from "../../csv-export/csv-builder.service";
 import { TranslateService } from "@ngx-translate/core";
+import { CsvExportService } from "../../csv-export/csv-export.service";
 
 @Component({
   selector: 'app-group-results',
@@ -89,8 +87,7 @@ export class SchoolResultsComponent implements OnInit {
               private filterOptionService: ExamFilterOptionsService,
               private schoolService: SchoolService,
               private angulartics2: Angulartics2,
-              private examFilterService: ExamFilterService,
-              private csvBuilder: CsvBuilder,
+              private csvExportService: CsvExportService,
               private translateService: TranslateService,
               public assessmentProvider: SchoolAssessmentService) {
   }
@@ -178,45 +175,11 @@ export class SchoolResultsComponent implements OnInit {
   }
 
   exportCsv(): void {
-    let assessmentExams: AssessmentExam[] = this.assessmentsComponent.assessmentExams;
-    let sourceData: any[] = [];
-    assessmentExams.forEach((assessmentExam: AssessmentExam) => {
-      let filteredExams: Exam[] = this.examFilterService.filterExams(assessmentExam, this.assessmentsComponent.clientFilterBy);
-      filteredExams.forEach((exam) => {
-        sourceData.push({
-          assessment: assessmentExam.assessment,
-          exam: exam
-        });
-      });
-    });
-
-    let getStudent = (item) => item.exam.student;
-    let getExam = (item) => item.exam;
-    let getAssessment = (item) => item.assessment;
-    let getIABExam = (item) => item.assessment.isIab ? item.exam : null;
-    let getNonIABExam = (item) => item.assessment.isIab ? null: item.exam;
-    let getNonIABMathExam = (item) => !item.assessment.isIab && item.assessment.subject === 'MATH' ? item.exam : null;
-    let getNonIABElaExam = (item) => !item.assessment.isIab && item.assessment.subject === 'ELA' ? item.exam : null;
-
-    let fileName: string = this._currentSchool.name +
+    let filename: string = this._currentSchool.name +
       "-" + this.translateService.instant(`labels.grades.${this._currentGrade.code}.short-name`) +
       "-" + new Date().toDateString();
 
-    this.csvBuilder
-      .newBuilder()
-      .withFilename(fileName)
-      .withStudent(getStudent)
-      .withExamDateAndSession(getExam)
-      .withAssessmentTypeNameAndSubject(getAssessment)
-      .withExamGradeAndStatus(getExam)
-      .withAchievementLevel(getNonIABExam)
-      .withReportingCategory(getIABExam)
-      .withScoreAndErrorBand(getExam)
-      .withMathClaimScores(getNonIABMathExam)
-      .withELAClaimScores(getNonIABElaExam)
-      .withGender(getStudent)
-      .withStudentContext(getExam)
-      .build(sourceData);
+    this.csvExportService.exportAssessmentExams(this.assessmentsComponent.assessmentExams, this.assessmentsComponent.clientFilterBy, filename);
   }
 
   private trackAnalyticsEvent(changedFilter: string) {
